requirements.txt
fastapi
uvicorn
pydantic
pytesseract
pdfplumber
spacy
fuzzywuzzy
python-Levenshtein
psycopg2-binary
sqlalchemy
langdetect
language-tool-python
Pillow